refactor: move config to global.config
